jwt をまとめる
