+++ src/repository.rs
@@ -4,9 +4,8 @@ use sqlx::{
     MySqlPool,
 };
 
-mod signup_jwt;
+mod jwt;
 pub mod users;
-mod users_jwt;
 mod users_session;
 
 #[derive(Clone)]
